+++ lib/cocoapods/sandbox/headers_store.rb
@@ -4,6 +4,8 @@ module Pod
     # the header search paths.
     #
     class HeadersStore
+      SEARCH_PATHS_KEY = Struct.new(:platform_name, :target_name)
+
       # @return [Pathname] the absolute path of this header directory.
       #
       def root
@@ -24,21 +26,32 @@ module Pod
         @sandbox       = sandbox
         @relative_path = relative_path
         @search_paths  = []
+        @search_paths_cache = {}
       end
 
       # @param  [Platform] platform
       #         the platform for which the header search paths should be
       #         returned
       #
+      # @param  [String] target_name
+      #         the target for which the header search paths should be
+      #         returned. This will return only header root scope e.g. `${PODS_ROOT}/Headers/Public`
+      #         if the target name specified is `nil`.
+      #
       # @return [Array<String>] All the search paths of the header directory in
       #         xcconfig format. The paths are specified relative to the pods
       #         root with the `${PODS_ROOT}` variable.
       #
-      def search_paths(platform)
-        platform_search_paths = @search_paths.select { |entry| entry[:platform] == platform.name }
-
+      def search_paths(platform, target_name = nil)
+        key = SEARCH_PATHS_KEY.new(platform.name, target_name)
+        return @search_paths_cache[key] if @search_paths_cache.key?(key)
+        platform_search_paths = @search_paths.select do |entry|
+          matches_platform = entry[:platform] == platform.name
+          matches_target = target_name.nil? || (entry[:path].basename.to_s == target_name)
+          matches_platform && matches_target
+        end
         headers_dir = root.relative_path_from(sandbox.root).dirname
-        ["${PODS_ROOT}/#{headers_dir}/#{@relative_path}"] + platform_search_paths.uniq.map { |entry| "${PODS_ROOT}/#{headers_dir}/#{entry[:path]}" }
+        @search_paths_cache[key] = ["${PODS_ROOT}/#{headers_dir}/#{@relative_path}"] + platform_search_paths.uniq.map { |entry| "${PODS_ROOT}/#{headers_dir}/#{entry[:path]}" }
       end
 
       # Removes the directory as it is regenerated from scratch during each